requirements-mac.txt
absl-py==1.4.0
asgiref==3.6.0
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
black==23.1.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
debugpy==1.6.6
decorator==5.1.1
Django==4.1.6
djangorestframework==3.14.0
executing==1.2.0
filelock==3.9.0
flatbuffers==23.1.21
gast==0.4.0
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.8.0
huggingface-hub==0.12.1
idna==3.4
importlib-metadata==6.0.0
ipykernel==6.21.2
ipython==8.10.0
jedi==0.18.2
jupyter_client==8.0.3
jupyter_core==5.2.0
keras==2.11.0
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
nest-asyncio==1.5.6
numpy==1.24.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
parso==0.8.3
pathspec==0.11.0
pickleshare==0.7.5
platformdirs==3.0.0
prompt-toolkit==3.0.37
protobuf==3.19.6
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.14.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
# pywin32==305
PyYAML==6.0
pyzmq==25.0.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
# tensorflow==2.11.0
tensorflow-estimator==2.11.0
# tensorflow-intel==2.11.0
# tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
tokenizers==0.13.2
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
transformers==4.26.1
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
wcwidth==0.2.6
Werkzeug==2.2.3
wrapt==1.14.1
zipp==3.14.0
